fixes issue 40: Here I renamed the alphafold3_pytorch/typing.py file to alphafold3_pytorch/custom_typing.py and fixed the imports in all files.
