+++ alphafold3_pytorch/alphafold3.py
@@ -19,7 +19,7 @@ from torch.nn import (
 
 from typing import List, Literal, Tuple, NamedTuple, Callable
 
-from alphafold3_pytorch.typing import (
+from alphafold3_pytorch.custom_typing import (
     Float,
     Int,
     Bool,
