+++ alphafold3_pytorch/trainer.py
@@ -8,7 +8,7 @@ from alphafold3_pytorch.attention import pad_at_dim
 
 from typing import TypedDict, List, Callable
 
-from alphafold3_pytorch.typing import (
+from alphafold3_pytorch.custom_typing import (
     typecheck,
     beartype_isinstance,
     Int, Bool, Float
